# fix float8 rowwise inference perf with torch.compile