In https://github.com/pytorch/ao/pull/2379, logic was added which prevented torchinductor from fusing the activation quantization for float8 inference.  Here are some logs which show the extra kernels being added by that PR to float8 inference on NVIDIA GPUs: https://www.internalfb.com/phabricator/paste/view/P1891592748 .

This PR reverts most of https://github.com/pytorch/ao/pull/2379, and adds a test to ensure we see the correct # of GPU kernels for float8 tensorwise and rowwise quantization.  We'll have to re-do https://github.com/pytorch/ao/pull/2379 without breaking this test.

Perf impact of this PR on MKN == 1024, 2048, 4096 on an NVIDIA H100 for float8 rowwise inference:
* torch.compile default: 50.9us -> 24.2us (2.1x speedup)
* torch.compile reduce-overhead: 66.4us -> 41.5us (1.6x speedup)
* logs: https://gist.github.com/vkuzo/1d35bea7fa5d38040d4d14d95627ce7f

Note that I added a benchmark to `benchmarks/inference/bench_float8_inference.py` to reproduce the numbers above, but I ran this benchmark out-of-tree to get the actual numbers, for easier comparison of before-this-PR vs after-this-PR.

Summary:

Test Plan:

```bash
TORCHINDUCTOR_FORCE_DISABLE_CACHES=1 pytest test/dtypes/test_affine_quantized_float.py -s -k expected_kernels_on_gpu
```

Reviewers:

Subscribers:

Tasks:

Tags: